State (> Revision): Fix using legacy pattern for selectors
